+++ src/main/java/depth/main/wishwesee/domain/auth/dto/request/RefreshTokenReq.java
@@ -1,4 +1,4 @@
-package depth.main.wishwesee.domain.auth.dto.req;
+package depth.main.wishwesee.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
