+++ src/main/java/depth/main/wishwesee/domain/auth/dto/response/AuthRes.java
@@ -1,4 +1,4 @@
-package depth.main.wishwesee.domain.auth.dto.res;
+package depth.main.wishwesee.domain.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
