+++ src/main/java/depth/main/wishwesee/domain/auth/dto/response/TokenMapping.java
@@ -1,4 +1,4 @@
-package depth.main.wishwesee.domain.auth.dto.res;
+package depth.main.wishwesee.domain.auth.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
