+++ src/main/java/depth/main/wishwesee/domain/invitation/dto/request/CreateFeedbackReq.java
@@ -1,4 +1,4 @@
-package depth.main.wishwesee.domain.invitation.dto.req;
+package depth.main.wishwesee.domain.invitation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
