+++ src/main/java/depth/main/wishwesee/domain/user/dto/response/UserProfileRes.java
@@ -1,4 +1,4 @@
-package depth.main.wishwesee.domain.user.dto.res;
+package depth.main.wishwesee.domain.user.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
